Switch to 2-digit checksum
The cardinal-31 scheme results in one substitution error, A/7,
which, while admittedly unlikely, does look bad on the fuzzy
testing. Solving single errors is something that most other
checksums do since Verhoeff; if we can’t clear that bar, it does not
bring confidence to onlookers, no matter the genuine risk.
Mitigations I initially envisioned turned out not to work (eg.
replacing 7 with AB).

On the other hand, the work we did for that scheme made it seem more
tractable to clear the cardinal-2, k=5 scheme, which I initially
discarded as too computationally intensive, because of the
necessary brute-forcing of various operations.

We will therefore separate the work in two: one two-character
checksum using a cardinal-1021 scheme similar to the cardinal-31
one, and one single-character checksum using the cardinal-2 scheme.