Refactor binaries and fuzzy tests
Refactor the base32check binary to rely on another file.
Add the base32check1 and base32check2 using that other file.

Refactor the gen-primitive binary to have essential functions
outside.

Refactor fuzzy tests to have essential functions outside.

That way, we can automate the computation of the parameters with the
highest detection rate.