Use a detection factor instead of a detection rate per bit
The detection rate per bit was computed as `detection rate / checksum bits`,
which implied that a given design had a detection rate grow linearly with the
number of bits. This obviously cannot be true, as the detection rate cannot
exceed 1.

Furthermore, because of the rule of diminishing returns, every additional bit
contributes less and less to the overall detection rate.

We therefore switched to an exponential-based growth formula converging to 1,
parameterized by what we call the detection factor. A higher detection factor
indicates that a design grows to 1 faster.